Extract all menu definitions
