+++ desktop/app.js
@@ -3,7 +3,6 @@
 const {
   app,
   BrowserWindow,
-  dialog,
   ipcMain,
   shell,
   Menu,
@@ -13,11 +12,8 @@ const {
 const path = require('path');
 const windowStateKeeper = require('electron-window-state');
 
-const buildViewMenu = require('./menus/view-menu');
-const buildEditMenu = require('./menus/edit-menu');
+const createMenuTemplate = require('./menus');
 const { isDev } = require('./env');
-
-const DialogTypes = require('../shared/dialog-types');
 
 require('module').globalPaths.push(path.resolve(path.join(__dirname)));
 
@@ -172,249 +168,3 @@ module.exports = function main() {
   app.on('ready', activateWindow);
   app.on('activate', activateWindow);
 };
-
-function createMenuTemplate(settings) {
-  const name = app.getName();
-
-  const aboutMenuItem = {
-    label: '&About ' + name,
-    click: function(item, focusedWindow) {
-      if (focusedWindow) {
-        focusedWindow.webContents.send('appCommand', {
-          action: 'showDialog',
-          dialog: DialogTypes.ABOUT,
-        });
-      }
-    },
-  };
-
-  const settingsMenuItem = {
-    label: 'P&references',
-    accelerator: 'CommandOrControl+,',
-    click: function(item, focusedWindow) {
-      if (focusedWindow) {
-        focusedWindow.webContents.send('appCommand', {
-          action: 'showDialog',
-          dialog: DialogTypes.SETTINGS,
-        });
-      }
-    },
-  };
-
-  const helpMenu = {
-    label: '&Help',
-    submenu: [
-      {
-        label: 'Help && &Support',
-        click: function() {
-          shell.openExternal('https://simplenote.com/help');
-        },
-      },
-    ],
-  };
-
-  const fileMenu = {
-    label: '&File',
-    submenu: [
-      {
-        label: '&New Note',
-        accelerator: 'CommandOrControl+N',
-        click(item, focusedWindow) {
-          if (focusedWindow) {
-            focusedWindow.webContents.send('appCommand', { action: 'newNote' });
-          }
-        },
-      },
-      {
-        type: 'separator',
-      },
-      {
-        label: '&Import Notes…',
-        click: function(item, focusedWindow) {
-          if (focusedWindow) {
-            focusedWindow.webContents.send('appCommand', {
-              action: 'showDialog',
-              dialog: DialogTypes.IMPORT,
-            });
-          }
-        },
-      },
-      {
-        label: '&Export Notes…',
-        accelerator: 'CommandOrControl+Shift+E',
-        click(item, focusedWindow) {
-          if (focusedWindow) {
-            dialog.showSaveDialog(
-              focusedWindow,
-              {
-                title: 'Save export as .zip archive',
-                defaultPath: path.join(app.getPath('desktop'), 'notes.zip'),
-              },
-              filename =>
-                focusedWindow.webContents.send('appCommand', {
-                  action: 'exportZipArchive',
-                  filename,
-                })
-            );
-          }
-        },
-      },
-      {
-        type: 'separator',
-      },
-      {
-        label: '&Print',
-        accelerator: 'CommandOrControl+P',
-        click(item, focusedWindow) {
-          if (focusedWindow) {
-            focusedWindow.webContents.send('appCommand', {
-              action: 'setShouldPrintNote',
-            });
-          }
-        },
-      },
-    ],
-  };
-
-  const windowMenu = {
-    label: '&Window',
-    role: 'window',
-    submenu: [
-      {
-        label: '&Minimize',
-        accelerator: 'CommandOrControl+M',
-        role: 'minimize',
-      },
-      {
-        label: '&Close',
-        accelerator: 'CommandOrControl+W',
-        role: 'close',
-      },
-    ],
-  };
-
-  // linux menu item adjustments
-  if (process.platform === 'linux') {
-    // add about menu item to Help
-    helpMenu['submenu'].push({
-      type: 'separator',
-    });
-    helpMenu['submenu'].push(aboutMenuItem);
-    helpMenu['submenu'].push({
-      label: 'Help',
-      accelerator: 'F1',
-      click: function() {
-        shell.openExternal('https://simplenote.com/help');
-      },
-    });
-    // add exit and settings items to File
-    fileMenu['submenu'].push({
-      type: 'separator',
-    });
-    fileMenu['submenu'].push(settingsMenuItem);
-    fileMenu['submenu'].push({
-      type: 'separator',
-    });
-    fileMenu['submenu'].push({
-      label: 'E&xit',
-      accelerator: 'Ctrl+Q',
-      click: function() {
-        app.quit();
-      },
-    });
-  }
-
-  // non-osx and non-linux (eg:windows, freebsd, sunos) menu item adjustments
-  if (process.platform !== 'darwin' && process.platform !== 'linux') {
-    // add about menu item to Help
-    helpMenu['submenu'].push({
-      type: 'separator',
-    });
-    helpMenu['submenu'].push(aboutMenuItem);
-
-    // add exit and settings items to File
-    fileMenu['submenu'].push({
-      type: 'separator',
-    });
-    fileMenu['submenu'].push(settingsMenuItem);
-    fileMenu['submenu'].push({
-      type: 'separator',
-    });
-    fileMenu['submenu'].push({
-      label: 'E&xit',
-      accelerator: 'Alt+F4',
-      click: function() {
-        app.quit();
-      },
-    });
-  }
-
-  const menuTemplate = [
-    fileMenu,
-    buildEditMenu(settings),
-    buildViewMenu(settings),
-    windowMenu,
-    helpMenu,
-  ];
-
-  if (process.platform === 'darwin') {
-    // Add the 'Simplenote' menu for os x
-    menuTemplate.unshift({
-      label: name,
-      submenu: [
-        aboutMenuItem,
-        {
-          type: 'separator',
-        },
-        settingsMenuItem,
-        {
-          type: 'separator',
-        },
-        {
-          label: 'Services',
-          role: 'services',
-          submenu: [],
-        },
-        {
-          type: 'separator',
-        },
-        {
-          label: 'Hide ' + name,
-          accelerator: 'CommandOrControl+H',
-          role: 'hide',
-        },
-        {
-          label: 'Hide Others',
-          accelerator: 'CommandOrControl+Shift+H',
-          role: 'hideothers',
-        },
-        {
-          label: 'Show All',
-          role: 'unhide',
-        },
-        {
-          type: 'separator',
-        },
-        {
-          label: 'Quit',
-          accelerator: 'CommandOrControl+Q',
-          click: function() {
-            app.quit();
-          },
-        },
-      ],
-    });
-
-    windowMenu.submenu.push(
-      {
-        type: 'separator',
-      },
-      {
-        label: 'Bring All to Front',
-        role: 'front',
-      }
-    );
-  }
-
-  return menuTemplate;
-}
