+++ desktop/menus/help-menu.js
@@ -1,0 +1,22 @@
+const { shell } = require('electron');
+
+const menuItems = require('./menu-items');
+const platform = require('../platform');
+
+const submenu = [
+  {
+    label: 'Help && &Support',
+    accelerator: platform.isLinux() ? 'F1' : null,
+    click: () => shell.openExternal('https://simplenote.com/help'),
+  },
+];
+
+const defaultSubmenuAdditions = [{ type: 'separator' }, menuItems.about];
+
+const helpMenu = {
+  label: '&Help',
+  role: 'help',
+  submenu: platform.isOSX() ? submenu : submenu.concat(defaultSubmenuAdditions),
+};
+
+module.exports = helpMenu;
