+++ desktop/menus/mac-app-menu.js
@@ -1,0 +1,25 @@
+const { app } = require('electron');
+
+const menuItems = require('./menu-items');
+
+const macAppMenu = {
+  label: app.getName(),
+  submenu: [
+    menuItems.about,
+    { type: 'separator' },
+    menuItems.preferences,
+    { type: 'separator' },
+    {
+      role: 'services',
+      submenu: [],
+    },
+    { type: 'separator' },
+    { role: 'hide' },
+    { role: 'hideothers' },
+    { role: 'unhide' },
+    { type: 'separator' },
+    { role: 'quit' },
+  ],
+};
+
+module.exports = macAppMenu;
